chore: refactor hnsw index into separate files
