+++ src/core/search/indices.h
@@ -191,28 +191,6 @@ struct FlatVectorIndex : public BaseVectorIndex {
   PMR_NS::vector<float> entries_;
 };
 
-struct HnswlibAdapter;
-class HnswVectorIndex {
- public:
-  explicit HnswVectorIndex(const search::SchemaField::VectorParams& params,
-                           PMR_NS::memory_resource* mr = PMR_NS::get_default_resource());
-
-  ~HnswVectorIndex();
-
-  bool Add(search::GlobalDocId id, const search::DocumentAccessor& doc, std::string_view field);
-  void Remove(search::GlobalDocId id, const search::DocumentAccessor& doc, std::string_view field);
-
-  std::vector<std::pair<float, GlobalDocId>> Knn(float* target, size_t k,
-                                                 std::optional<size_t> ef) const;
-  std::vector<std::pair<float, GlobalDocId>> Knn(float* target, size_t k, std::optional<size_t> ef,
-                                                 const std::vector<GlobalDocId>& allowed) const;
-
- private:
-  size_t dim_;
-  VectorSimilarity sim_;
-  std::unique_ptr<HnswlibAdapter> adapter_;
-};
-
 struct GeoIndex : public BaseIndex {
   using point =
       boost::geometry::model::point<double, 2,
