+++ src/server/search/global_hnsw_index.h
@@ -13,7 +13,7 @@
 #include <vector>
 
 #include "core/search/base.h"
-#include "core/search/indices.h"
+#include "core/search/hnsw_index.h"
 #include "core/search/search.h"
 #include "server/search/doc_index.h"
 #include "server/tx_base.h"
